Refactor JobWorkEnv includes: formik update, renaming files, radiogroup focing on error issue.
